requirements.txt
discord.py
python-dotenv
requests
psycopg2-binary
pinecone-client
supabase
google-cloud-aiplatform
openai
anthropic
groq
livekit
n8n-sdk-python
linear-python
arcadepy
brave-search
firecrawl-py
stripe
elevenlabs
deepgram-sdk
pyyaml
pywin32
